+++ app/src/main/kotlin/com/algorand/android/ui/compose/widget/AddressCard.kt
@@ -213,83 +213,3 @@ fun PeraIconHighlightedText(modifier: Modifier = Modifier, text: String) {
     }
 }
 
-@PreviewLightDark
-@Composable
-fun PeraIconHighlightedTextPreview() {
-    PeraTheme {
-        PeraIconHighlightedText(
-            modifier = Modifier.padding(start = 37.dp),
-            text = stringResource(R.string.scan_new_addresses)
-        )
-    }
-}
-
-@PreviewLightDark
-@Composable
-fun StandardAccountPreview() {
-    PeraTheme {
-        Column {
-            AddressCard(
-                modifier = Modifier,
-                title = "CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5",
-                subTitle = null,
-                accountRegistrationType = AccountRegistrationType.Algo25,
-                accountRekeyRegistrationType = null,
-                icon = R.drawable.ic_wallet,
-                iconBackgroundColor = PeraTheme.colors.wallet.wallet4.background,
-                onAccountItemButtonClick = { },
-            )
-        }
-    }
-}
-
-@PreviewLightDark
-@Composable
-fun HDAccountPreview() {
-    PeraTheme {
-        AddressCard(
-            modifier = Modifier,
-            title = "HD Wallet #1",
-            subTitle = "CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5",
-            accountRegistrationType = AccountRegistrationType.HdKey,
-            accountRekeyRegistrationType = null,
-            icon = R.drawable.ic_wallet_address,
-            iconBackgroundColor = PeraTheme.colors.wallet.wallet4.background,
-            onAccountItemButtonClick = { },
-        )
-    }
-}
-
-@PreviewLightDark
-@Composable
-fun StandardRekeyPreview() {
-    PeraTheme {
-        AddressCard(
-            modifier = Modifier,
-            title = "CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5",
-            subTitle = null,
-            accountRegistrationType = AccountRegistrationType.Algo25,
-            accountRekeyRegistrationType = AccountRegistrationType.Algo25,
-            icon = R.drawable.ic_wallet,
-            iconBackgroundColor = PeraTheme.colors.wallet.wallet4.background,
-            onAccountItemButtonClick = { },
-        )
-    }
-}
-
-@PreviewLightDark
-@Composable
-fun HdRekeyPreview() {
-    PeraTheme {
-        AddressCard(
-            modifier = Modifier,
-            title = "HD Wallet #1",
-            subTitle = "CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5CJR5",
-            accountRegistrationType = AccountRegistrationType.HdKey,
-            accountRekeyRegistrationType = AccountRegistrationType.Algo25,
-            icon = R.drawable.ic_wallet,
-            iconBackgroundColor = PeraTheme.colors.wallet.wallet4.background,
-            onAccountItemButtonClick = { },
-        )
-    }
-}
